requirements.txt
amqp==5.3.1
anyio==4.6.2.post1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
deprecation==2.1.0
Django==5.1.3
django-colorfield==0.11.0
django-debug-toolbar==4.4.6
django-email-verification==0.3.3
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
kombu==5.4.2
mysqlclient==2.2.4
novaposhta-python-client==0.1.10
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.48
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
validators==0.34.0
vine==5.1.0
wcwidth==0.2.13
